add docs to kapi and move system call wrappers there
