+++ kapi/Cargo.toml
@@ -8,4 +8,8 @@ edition = "2021"
 [dependencies]
 serde = { version = "1.0", features = ["derive"], default-features = false }
 snafu = { version = "0.8", default-features = false, features = ["unstable-core-error", "rust_1_65"] }
+log = "0.4"
 bytemuck = "1"
+
+[features]
+kernel = []
